Extract DSPy event log callback
